requirements.txt
moviepy
faster-whisper
python-dotenv
pytube
Pillow
langchain-openai
langchain-core
python-telegram-bot
langchain
selenium
aiofiles
yt-dlp
loguru